make image module cleaner
